Move from `cache.guava` to `guava.cache`
